requirements.txt
Flask==3.0.2
numpy==1.26.4
tensorflow-cpu==2.15.0
keras==2.15.0
opencv-python-headless
gunicorn==21.2.0